+++ ui/packages/shared/profile/src/ProfileIcicleGraph/IcicleGraphArrow/useNodeColor.ts
@@ -12,7 +12,7 @@
 // limitations under the License.
 
 import {EVERYTHING_ELSE} from '@parca/store';
-import {diffColor, getLastItem} from '@parca/utilities';
+import {diffColor, diffColorPerSecond, getLastItem} from '@parca/utilities';
 
 interface mappingColors {
   [key: string]: string;
@@ -22,7 +22,9 @@ interface Props {
   isDarkMode: boolean;
   compareMode: boolean;
   cumulative: bigint;
+  cumulativePerSecond: number | null;
   diff: bigint | null;
+  diffPerSecond: number | null;
   mappingColors: mappingColors;
   functionName: string | null;
   mappingFile: string | null;
@@ -32,12 +34,18 @@ const useNodeColor = ({
   isDarkMode,
   compareMode,
   cumulative,
+  cumulativePerSecond,
   diff,
+  diffPerSecond,
   mappingColors,
   functionName,
   mappingFile,
 }: Props): string => {
   if (compareMode) {
+    if (cumulativePerSecond !== null && diffPerSecond !== null) {
+      return diffColorPerSecond(diffPerSecond, cumulativePerSecond, isDarkMode);
+    }
+
     return diffColor(diff ?? 0n, cumulative, isDarkMode);
   }