+++ ui/packages/app/web/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.657](https://github.com/parca-dev/parca/compare/@parca/web@0.16.656...@parca/web@0.16.657) (2024-04-26)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.656](https://github.com/parca-dev/parca/compare/@parca/web@0.16.655...@parca/web@0.16.656) (2024-04-25)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.655](https://github.com/parca-dev/parca/compare/@parca/web@0.16.654...@parca/web@0.16.655) (2024-04-24)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.654](https://github.com/parca-dev/parca/compare/@parca/web@0.16.653...@parca/web@0.16.654) (2024-04-24)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.653](https://github.com/parca-dev/parca/compare/@parca/web@0.16.652...@parca/web@0.16.653) (2024-04-23)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.652](https://github.com/parca-dev/parca/compare/@parca/web@0.16.651...@parca/web@0.16.652) (2024-04-20)
+
+**Note:** Version bump only for package @parca/web
+
+## [0.16.651](https://github.com/parca-dev/parca/compare/@parca/web@0.16.650...@parca/web@0.16.651) (2024-04-20)
+
+**Note:** Version bump only for package @parca/web
+
 ## [0.16.650](https://github.com/parca-dev/parca/compare/@parca/web@0.16.649...@parca/web@0.16.650) (2024-04-16)
 
 **Note:** Version bump only for package @parca/web