+++ ui/packages/shared/utilities/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.65](https://github.com/parca-dev/parca/compare/@parca/utilities@0.0.64...@parca/utilities@0.0.65) (2024-04-26)
+
+**Note:** Version bump only for package @parca/utilities
+
+## 0.0.64 (2024-04-25)
+
+**Note:** Version bump only for package @parca/utilities
+
 ## 0.0.63 (2024-03-31)
 
 **Note:** Version bump only for package @parca/utilities